Main xcy

+++ ProjectHelper/items/users/views.py
@@ -1068,6 +1068,7 @@ class ChangeHeadImage(View):
             print('avatar exception')
             return JsonResponse({"ChangeHeadImage": "failed"})
 
+
 class ShowHeadImage(View):
 
     def get(self, request):
@@ -1624,19 +1625,29 @@ class TeacherKickMember(View):
         :return:
         """
         try:
-            # TODO: 判断学生是否为队长， 是则转移权限
             group_id = eval(request.body.decode()).get("group_id")
             token = eval(request.body.decode()).get("token")
             student_id = get_sid(token)
             target_id = eval(request.body.decode()).get("t_sid")
             user = UserProfile.objects.get(student_id=target_id)
-            # TODO:等待权限判断，能否给course_id
-            # auth = Authority.objects.get(user_id=student_id, type="teach", course_id=course_id)
-            # if auth.end_time > datetime.datetime.now() > auth.start_time:
-            group = GroupOrg.objects.get(group_name_id=group_id)
-            GroupOrg.objects.filter(group_name_id=group_id).update(member=group.member - 1)
-            UserGroup.objects.filter(group_name_id=group_id, user_name_id=user.id).delete()
-            return JsonResponse({"TeacherKickMemberCheck": "success"})
+            group = GroupOrg.objects.get(id=group_id)
+            project = Project.objects.get(id=group.project_id)
+            course = Course.objects.get(id=project.course_id)
+            auth = Authority.objects.get(user_id=student_id, type="group", course_id=course.id)
+            if auth.end_time > datetime.datetime.now() > auth.start_time:
+                if group.member == 1:
+                    GroupOrg.objects.filter(id=group_id).delete()
+                else:
+                    GroupOrg.objects.filter(id=group_id).update(member=group.member - 1)
+                if user.id == group.captain_name_id:
+                    member = UserGroup.objects.filter(group_name_id=group_id)
+                    for i in member:
+                        if i.user_name_id != user.id:
+                            GroupOrg.objects.filter(id=group_id).update(captain_name_id=i.user_name_id)
+                            break
+                UserGroup.objects.filter(group_name_id=group_id, user_name_id=user.id).delete()
+                return JsonResponse({"TeacherKickMemberCheck": "success"})
+            return JsonResponse({"TeacherKickMemberCheck": "no auth"})
         except Exception as e:
             logger.debug('%s %s', self, e)
             return JsonResponse({"TeacherKickMemberCheck": "failed"})
@@ -2328,6 +2339,12 @@ class SubmitEvent(View):
             project = Project.objects.get(id=event.project_id)
             course_id = project.course_id
             course = Authority.objects.get(user_id=user_id, type="eventValid", course_id=course_id)
+            user_group = UserGroup.objects.filter(user_name_id=user_id)
+            group = None
+            for i in user_group:
+                group = GroupOrg.objects.get(id=i.group_name_id)
+                if group.project_id == project.id:
+                    break
             if course.end_time > now > course.start_time:
                 if event.type == "choose":
                     pass
@@ -2347,17 +2364,16 @@ class SubmitEvent(View):
                         else:
                             if parameter['options'][i][1] == 0:
                                 raise
-                    ParticipantEvent.objects.filter(event_id_id=event_id, user_id=user_id).delete()
+                    ParticipantEvent.objects.filter(event_id_id=event_id, group_id=group.id).delete()
                     for i in data:
                         if parameter['partitionType'] == 'timeSlot':
                             start_time = datetime.datetime.fromtimestamp(parameter['options'][i][0] // 1000)
                             end_time = datetime.datetime.fromtimestamp(parameter['options'][i][1] // 1000)
-                            ParticipantEvent.objects.create(event_id_id=event_id, user_id=user_id,
+                            ParticipantEvent.objects.create(event_id_id=event_id, group_id=group.id,
                                                             start_time=start_time, end_time=end_time)
                             parameter['options'][i][2] -= 1
                         else:
-                            choice = parameter['options'][i][0]
-                            ChooseEvent.objects.create(event_id_id=event_id, user_id=user_id, choice=choice)
+                            ChooseEvent.objects.create(event_id_id=event_id, group_id=group.id, choice=i)
                             parameter['options'][i][1] -= 1
                 return JsonResponse({"SubmitEvent": "success"})
             return JsonResponse({"SubmitEvent": "failed"})
@@ -2393,16 +2409,18 @@ class GetEventDetail(View):
                 events = {'event_type': event.type, 'event_title': event.title,
                           'event_detail': json.loads(event.parameter), 'introduction': event.detail,
                           'publisher': publisher.student_id}
+                isStudent = True
                 if auth.count() != 0:
                     for k in auth:
                         if k.end_time > datetime.datetime.now() > k.start_time:
+                            isStudent = False
                             events['data'] = []
                             if event.type == "choose":
                                 choices = ChooseEvent.objects.filter(event_id_id=event.id)
                                 for j in choices:
-                                    student = UserProfile.objects.get(id=j.user_id)
-                                    events['data'].append({'choice': j.choice, 'student_id': student.student_id,
-                                                           'student_name': student.real_name})
+                                    group = GroupOrg.objects.get(id=j.group_id)
+                                    events['data'].append({'choice': j.choice, 'group_id': j.group_id,
+                                                           'group_name': group.name})
                             elif event.type == "attachment":
                                 choices = ProjectAttachment.objects.filter(event_id=event.id)
                                 for j in choices:
@@ -2412,13 +2430,86 @@ class GetEventDetail(View):
                             elif event.type == "partition":
                                 choices = ParticipantEvent.objects.filter(event_id_id=event.id)
                                 for j in choices:
-                                    student = UserProfile.objects.get(id=j.user_id)
+                                    group = GroupOrg.objects.get(id=j.group_id)
                                     events['data'].append({'start_time': j.start_time, 'end_time': j.end_time,
-                                                           'student_id': student.student_id,
-                                                           'student_name': student.real_name})
+                                                           'group_id': j.group_id, 'group_name': group.name})
                             break
+                    if isStudent:
+                        if event.type == "choose":
+                            choices = ChooseEvent.objects.filter(event_id_id=event.id)
+                            for j in choices:
+                                group = GroupOrg.objects.get(id=j.group_id)
+                                events['data'] = {'choice': j.choice, 'group_id': j.group_id, 'group_name': group.name}
+                        elif event.type == "attachment":
+                            choices = ProjectAttachment.objects.filter(event_id=event.id)
+                            for j in choices:
+                                group = GroupOrg.objects.get(id=j.group_id)
+                                events['data'] = {'path': j.file_path, 'group_id': j.group_id,
+                                                  'group_name': group.name}
+                        elif event.type == "partition":
+                            choices = ParticipantEvent.objects.filter(event_id_id=event.id)
+                            for j in choices:
+                                group = GroupOrg.objects.get(id=j.group_id)
+                                events['data'] = {'start_time': j.start_time, 'end_time': j.end_time,
+                                                  'group_id': j.group_id, 'group_name': group.name}
+
                 return JsonResponse({"Data": events, "GetEventDetail": "success"})
             return JsonResponse({"GetEventDetail": "no auth"})
+
+        except Exception as e:
+            logger.debug('%s %s', self, e)
+            return JsonResponse({"GetEventDetail": "failed"})
+
+
+class GetAllPartition(View):
+    def post(self, request):
+        """
+        user with "eventEdit" authority can get all event partition
+        :param token: token
+                project_id: id of project
+        :return:
+        """
+        try:
+            token = eval(request.body.decode()).get("token")
+            student_id = get_sid(token)
+            project_id = eval(request.body.decode()).get("project_id")
+            user = UserProfile.objects.get(student_id=student_id)
+            user_id = user.id
+            project = Project.objects.get(id=project_id)
+            course_id = project.course_id
+            auth = Authority.objects.get(user_id=user_id, type="eventEdit", course_id=course_id)
+            partitions = []
+            event = {}
+
+            if auth.end_time > datetime.datetime.now() > auth.start_time:
+                events = Event.objects.filter(project_id=project_id)
+                for i in events:
+                    if i.type == "partition":
+                        event = {'id': i.id, 'event_title': i.title}
+                        event_detail = json.loads(i.parameter)
+                        if event_detail['partitionType'] == 'normal':
+                            choices = ChooseEvent.objects.filter(event_id_id=i.id)
+                            choice = {}
+                            for j in range(len(event_detail['options'])):
+                                choice[event_detail['options'][j][0]] = []
+                            for j in choices:
+                                group = GroupOrg.objects.get(id=j.group_id)
+                                choice[event_detail['options'][int(j)][0]].append({'group_id': group.id,
+                                                                                   'group_name': group.name})
+                        else:
+                            choices = ParticipantEvent.objects.filter(event_id_id=i.id)
+                            choice = {}
+                            for j in range(len(event_detail['options'])):
+                                string = event_detail['options'][j][0] + '-' + event_detail['options'][j][1]
+                                choice[string] = []
+                            for j in choices:
+                                group = GroupOrg.objects.get(id=j.group_id)
+                                choice[event_detail['options'][int(j)][0]].append({'group_id': group.id,
+                                                                                   'group_name': group.name})
+                        event['data'] = choice
+                    partitions.append(event)
+
+            return JsonResponse({"Data": partitions, "GetEventDetail": "no auth"})
 
         except Exception as e:
             logger.debug('%s %s', self, e)
